# internal/api/openapi.vohub.yaml
openapi: 3.0.3
info:
  title: voHub Main API
  version: 1.0.0
  description: |
    voHub 主服务控制面 API。
    - 服务器前缀：`/api`
    - 登录方式：`POST /auth/login`
    - 受保护接口统一使用 `Authorization: Bearer <token>`
    - 文档页面：`GET /docs`
servers:
  - url: /api
    description: voHub main API prefix
security:
  - BearerAuth: []
...
paths:
...
  /openapi.yaml:
    get:
      tags: [settings]
      summary: 获取 OpenAPI YAML
      responses:
        '200':
          description: OpenAPI YAML 文档
          content:
            application/yaml:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /openapi.json:
    get:
      tags: [settings]
      summary: 获取 OpenAPI JSON
      responses:
        '200':
          description: OpenAPI JSON 文档
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /websheets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    get:
      tags: [vowifi]
      summary: 获取 Websheet 引导页
      security: []
      responses:
        '200': {description: Websheet HTML}
  /websheets/{id}/proxy:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    get: &websheetProxyOperation
      tags: [vowifi]
      summary: 代理 Websheet 请求
      security: []
      responses:
        '200': {description: 上游响应}
    post: *websheetProxyOperation
    put: *websheetProxyOperation
    patch: *websheetProxyOperation
    delete: *websheetProxyOperation
...
  /websheets/{id}/callback:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    post:
      tags: [vowifi]
      summary: 转发 Websheet 回调
      security: []
      responses:
        '200': {description: 回调结果}
  /websheets/{id}/done:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    post:
      tags: [vowifi]
      summary: 完成 Websheet 会话
      security: []
      responses:
        '200': {description: 已完成}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
...
  responses:
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        error:
          type: string
        warning:
          type: string
        result:
          type: object
          additionalProperties: true
        busy:
          type: boolean
...